+++ packages/styles/src/index.ts
@@ -13,5 +13,7 @@ export { tag } from "./packages/tag";
 export { text } from "./packages/text";
 export { title } from "./packages/title";
 export { toast } from "./packages/toast";
+export { list } from "./packages/list";
 export { toggle } from "./packages/toggle";
 export { link } from "./packages/link";
+export { tooltip } from "./packages/tooltip";