+++ packages/styles/src/packages/label/index.ts
@@ -1,4 +1,4 @@
-import { style } from "./label.style.css";
+import * as style from "./label.style.css";
 
 export const label = {
   style,