+++ packages/styles/webpack.config.js
@@ -25,8 +25,10 @@ module.exports = {
     "./packages/text/index": "./src/packages/text/index",
     "./packages/title/index": "./src/packages/title/index",
     "./packages/toast/index": "./src/packages/toast/index",
+    "./packages/list/index": "./src/packages/list/index",
     "./packages/toggle/index": "./src/packages/toggle/index",
     "./packages/link/index": "./src/packages/link/index",
+    "./packages/tooltip/index": "./src/packages/tooltip/index",
   },
   output: {
     path: path.resolve(__dirname, "dist"),