+++ packages/styles/CHANGELOG.md
@@ -8,6 +8,18 @@ Nimbus Styles deprive all styles needed to build nimbus components.
 
 - Added new style pack for link component. ([#35](https://github.com/TiendaNube/nimbus-design-system/pull/35) by [@juanchigallego](https://github.com/juanchigallego))
 
+## 2022-10-01
+
+### 🎉 New features
+
+- Added new style pack for tooltip component. ([#34](https://github.com/TiendaNube/nimbus-design-system/pull/34) by [@juniorconquista](https://github.com/juniorconquista))
+
+## 2022-09-29
+
+### 🎉 New features
+
+- Added new style pack for list component. ([#33](https://github.com/TiendaNube/nimbus-design-system/pull/33) by [@juanchigallego](https://github.com/juanchigallego))
+
 ## 2022-09-28
 
 ### 🎉 New features